apps/api/requirements.txt
fastapi
uvicorn[standard]
sqlmodel
psycopg2-binary
python-dotenv
pydantic-settings
fastapi-csrf-protect
passlib[bcrypt]
python-jose[cryptography]
python-multipart
